Hospital-Management-System
It is the user friendly application for Hospital which reduces the burden and helps to manage all sections of hospital like reception, lab, inpatient/outpatient management and Billing etc., which improve the processing efficiency. It deals with the automating tasks of maintaining of laboratory And Diagnosis. In Hospital, Billing management is the key process.
The Website is for the automation of Hospital Management. 
It maintains two levels of users
	Administrator Level  
	User Level